% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{extract_sample_names_from_mutalisk_files}
\alias{extract_sample_names_from_mutalisk_files}
\title{Sample Names From Mutalisk File Contents}
\usage{
extract_sample_names_from_mutalisk_files(mutalisk_filenames)
}
\arguments{
\item{mutalisk_filenames}{names of mutalisk output files (character)}
}
\value{
sample name (string)
}
\description{
Sample Names From Mutalisk File Contents
}
