% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{extractSampleNamesFromMutaliskFilenames}
\alias{extractSampleNamesFromMutaliskFilenames}
\title{Sample Names From Mutalisk Output}
\usage{
extractSampleNamesFromMutaliskFilenames(mutalisk_filenames)
}
\arguments{
\item{mutalisk_filenames}{names of mutalisk output files (character)}
}
\value{
sample name (string)
}
\description{
Sample Names From Mutalisk Output
}
