% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutalisk_dataframe_inform_user_of_metadata}
\alias{mutalisk_dataframe_inform_user_of_metadata}
\title{Mutalisk Dataframe}
\usage{
mutalisk_dataframe_inform_user_of_metadata(mutalisk_dataframe, metadata)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe produced using mutalisk_to_dataframe}

\item{metadata}{Either a path to csv file OR a dataframe. Must contain a header line which contains a SampleID column that matches that of mutalisk_dataframe (string)}
}
\value{
mutalisk dataframe with metadata columns (data.frame)
}
\description{
Adds metadata from a file to the mutalisk dataframe
}
