% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutalisk_to_dataframe}
\alias{mutalisk_to_dataframe}
\title{Mutalisk files to dataframe}
\usage{
mutalisk_to_dataframe(mutalisk_files, sample_names_from_file_contents = FALSE)
}
\arguments{
\item{mutalisk_files}{a vector of filepaths, each leading to the report.txt files output when downloading best_signature results for all vcfs in cohort}

\item{sample_names_from_file_contents}{guess sample names from filecontents instead of filenames (flag)}
}
\value{
a dataframe containing three columns:
\enumerate{
\item \strong{SampleID}: a sample identifier.
\item \strong{Signatures}: an identifier for a particular signature.
\item \strong{Contributions}: the percentage contribution of the signature to the patients genetic profile (0.1 = 10 percent).
}
}
\description{
Mutalisk files to dataframe
}
