% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutalisk_to_dataframe_single_sample}
\alias{mutalisk_to_dataframe_single_sample}
\title{Mutalisk to dataframe}
\usage{
mutalisk_to_dataframe_single_sample(mutalisk_file)
}
\arguments{
\item{mutalisk_file}{a vector of filepaths, each leading to the report.txt files output when downloading best_signature results for all vcfs in cohort}
}
\value{
tibble
}
\description{
You probably want to use tags$strong{mutalisk_to_dataframe} instead.
See ?mutalisk_to_dataframe
}
