% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutaliskBestSigatureDirectoryToDataframe}
\alias{mutaliskBestSigatureDirectoryToDataframe}
\title{Mutalisk directory to dataframe}
\usage{
mutaliskBestSigatureDirectoryToDataframe(directory)
}
\arguments{
\item{directory}{path to \strong{mutalisk_best_fit} folder.
To obtain, run your VCFs through mutalisk. Select Mutational Signature (Best only) and click 'Get the selected result for all samples at once'.
Then unzip the file, and youre ready to go}
}
\value{
tibble
}
\description{
Mutalisk directory to dataframe
}
