% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutaliskDataframeExpand}
\alias{mutaliskDataframeExpand}
\title{Expand mutalisk_dataframe}
\usage{
mutaliskDataframeExpand(mutalisk_dataframe)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe produced using mutaliskToDataFrame}
}
\value{
dataframe containing all combinations of Sample ID and Signatures.
For cases where a signature was not included in the 'best fit' subset, Contribution is set to 0\%.
}
\description{
In normal mutalisk dataframe, each sample has data for ONLY the 1-7 signatures that comprise the 'best fit'.
Bascically, we end up with a dataframe where not all signatures have entries for all samples. We call this 'implicit' missing values.
This means that a signature level jitterplot won't show the samples where it was not included in this 'best fit' set, when we'd actually want to know that it contributed 0\% to that sample.
This function fixes the issue by adding entries for ALL signature - sample pairs, with Contributions set to 0\% where relevant.
}
