% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutaliskToDataFrameSingleSample}
\alias{mutaliskToDataFrameSingleSample}
\title{Mutalisk to dataframe}
\usage{
mutaliskToDataFrameSingleSample(mutalisk_files)
}
\arguments{
\item{mutalisk_files}{a vector of filepaths, each leading to the report.txt files output when downloading best_signature results for all vcfs in cohort}
}
\value{
tibble
}
\description{
You probably want to use tags$strong{mutaliskToDataFrame} instead.
See ?mutaliskToDataFrame
}
