% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{pallette_cols23_customised}
\alias{pallette_cols23_customised}
\title{Custom Palette}
\usage{
pallette_cols23_customised(n)
}
\arguments{
\item{n}{number of colors in pallete to return}
}
\value{
character vector of colors
}
\description{
Custom Palette
}
\examples{
pallette_cols23_customised(2)
}
