% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{plot_signature_contribution_jitterplot}
\alias{plot_signature_contribution_jitterplot}
\title{Plot Signature-Level Dotplot
Plots a signature-Level dotplot}
\usage{
plot_signature_contribution_jitterplot(mutalisk_dataframe)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe that can be produced using mutalisk_best_signature_directory_to_dataframe. Can also just make it yourself, if you want to visualise non-mutalisk data. Dataframe just needs 3 columns:
\enumerate{
\item \strong{SampleID}: a sample identifier.
\item \strong{Signatures}: an identifier for a particular signature.
\item \strong{Contributions}: the percentage contribution of the signature to the patients genetic profile (0.1 = 10\%).
}}
}
\value{
a ggplot object
}
\description{
Plot Signature-Level Dotplot
Plots a signature-Level dotplot
}
