% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{plot_stacked_bar}
\alias{plot_stacked_bar}
\title{Cohort-Level Mutational Signature Visualisation}
\usage{
plot_stacked_bar(
  mutalisk_dataframe,
  lump_type = "min_prop",
  lump_min = 0.1,
  topn = 5,
  legend = "right",
  legend_direction = NA,
  pal = pals::kovesi.diverging_rainbow_bgymr_45_85_c67,
  color_of_other = "grey60",
  facet_column = NA,
  fontsize_strip = 18,
  fontsize_axis_title = 18
)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe that can be produced using mutalisk_best_signature_directory_to_dataframe. Can also just make it yourself, if you want to visualise non-mutalisk data. Dataframe just needs 3 columns:
\enumerate{
\item \strong{SampleID}: a sample identifier.
\item \strong{Signatures}: an identifier for a particular signature.
\item \strong{Contributions}: the percentage contribution of the signature to the patients genetic profile (0.1 = 10\%).
}}

\item{lump_type}{one of "min_prop", "topn", or "none".

\strong{min_prop} will allow lump together all signatures that contribute less than \strong{lump_min}.
\strong{topn} will keep the topn contributing signatures distinct, and lump the rest together (string)}

\item{lump_min}{see \strong{lump_type}}

\item{topn}{see \strong{lump_type}}

\item{legend}{Where should the legend be placed? ("top", "left", "bottom", "right")}

\item{legend_direction}{How should the legend be oriented. By defualt will guess based on position of legend ("vertical", "horizontal")}

\item{pal}{Palette to use for generating colours.}

\item{color_of_other}{colour of signatures lumped into 'other' (string)}

\item{facet_column}{name of column to use for faceting (string)}

\item{fontsize_strip}{fontsize of facet titles (number)}

\item{fontsize_axis_title}{fontsize of axis titles (number)}
}
\value{
a ggplot (gg)
}
\description{
A note of warning: for different mutalisk runs, this function will not enforce uniform colours for a single mutational signature. Better to get all data in at once and add a facet_wrap call
}
