% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{plot_stacked_bar_interactive}
\alias{plot_stacked_bar_interactive}
\title{Cohort-Level Mutational Signature Visualisation}
\usage{
plot_stacked_bar_interactive(
  mutalisk_dataframe,
  lump_type = "min_prop",
  lump_min = 0.1,
  topn = 5,
  legend = "top"
)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe produced using mutalisk_to_dataframe}

\item{lump_type}{one of "min_prop", "topn", or "none".

\strong{min_prop} will allow lump together all signatures that contribute less than \strong{lump_min}.
\strong{topn} will keep the topn contributing signatures distinct, and lump the rest together (string)}

\item{lump_min}{see \strong{lump_type}}

\item{topn}{see \strong{lump_type}}

\item{legend}{Where should the legend be placed? ("top", "left", "bottom", "right")}
}
\value{
plotly visualisation
}
\description{
A note of warning: for different mutalisk runs, this function will not enforce uniform colours for a single mutational signature. Better to get all data in at once and add a facet_wrap call
}
