% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{plotDistributionOfSignatureContributions}
\alias{plotDistributionOfSignatureContributions}
\title{Plot Signature-Level Dotplot
Plots a signature-Level dotplot}
\usage{
plotDistributionOfSignatureContributions(mutalisk_dataframe)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe produced using mutaliskToDataFrame}
}
\value{
a ggplot object
}
\description{
Plot Signature-Level Dotplot
Plots a signature-Level dotplot
}
