% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{plotStackedBar}
\alias{plotStackedBar}
\title{Cohort-Level Mutational Signature Visualisation}
\usage{
plotStackedBar(
  mutalisk_dataframe,
  lump_type = "min_prop",
  lump_min = 0.1,
  topn = 5,
  legend = "right",
  legend_direction = NA,
  pal = pals::kovesi.diverging_rainbow_bgymr_45_85_c67,
  facet_column = NA,
  facet_ncol = 1,
  fontsize_strip = 18,
  fontsize_axis_title = 18
)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe produced using mutaliskToDataFrame}

\item{lump_type}{one of "min_prop", "topn", or "none".

\strong{min_prop} will allow lump together all signatures that contribute less than \strong{lump_min}.
\strong{topn} will keep the topn contributing signatures distinct, and lump the rest together (string)}

\item{lump_min}{see \strong{lump_type}}

\item{topn}{see \strong{lump_type}}

\item{legend}{Where should the legend be placed? ("top", "left", "bottom", "right")}

\item{legend_direction}{How should the legend be oriented. By defualt will guess based on position of legend ("vertical", "horizontal")}

\item{pal}{Palette to use for generating colours.}

\item{facet_column}{name of column to use for faceting (string)}

\item{facet_ncol}{number of columns in faceted plot (number)}

\item{fontsize_strip}{fontsize of facet titles (number)}

\item{fontsize_axis_title}{fontsize of axis titles (number)}
}
\value{
a ggplot (gg)
}
\description{
A note of warning: for different mutalisk runs, this function will not enforce uniform colours for a single mutational signature. Better to get all data in at once and add a facet_wrap call
}
