% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutalisk_dataframe_add_metadata}
\alias{mutalisk_dataframe_add_metadata}
\title{Add sample metadata}
\usage{
mutalisk_dataframe_add_metadata(mutalisk_dataframe, sample_metadata)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe produced using mutalisk_to_dataframe}

\item{sample_metadata}{a dataframe containing a SampleID column and additional columns for each property you want to add as metadata (data.frame)}
}
\value{
mutalisk dataframe with additional metadata columns (data.frame)
}
\description{
Add sample metadata to mutalisk
}
