% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutalisk_dataframe_inform_user_of_metadata}
\alias{mutalisk_dataframe_inform_user_of_metadata}
\title{Mutalisk Dataframe}
\usage{
mutalisk_dataframe_inform_user_of_metadata(mutalisk_dataframe, metadata_file)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe that can be produced using mutalisk_best_signature_directory_to_dataframe. Can also just make it yourself, if you want to visualise non-mutalisk data. Dataframe just needs 3 columns:
\enumerate{
\item \strong{SampleID}: a sample identifier.
\item \strong{Signatures}: an identifier for a particular signature.
\item \strong{Contributions}: the percentage contribution of the signature to the patients genetic profile (0.1 = 10\%).
}}

\item{metadata_file}{path to csv or tsv file. Must contain a header line which contains a SampleID column (or Tumor_Sample_Barcode) that matches that of mutalisk_dataframe (string)}
}
\value{
mutalisk dataframe with metadata columns (data.frame)
}
\description{
Adds metadata from a file to the mutalisk dataframe
}
