% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseMutaliskResults.R
\name{mutalisk_dataframe_metadata_column_names}
\alias{mutalisk_dataframe_metadata_column_names}
\title{Mutalisk Dataframe}
\usage{
mutalisk_dataframe_metadata_column_names(mutalisk_dataframe)
}
\arguments{
\item{mutalisk_dataframe}{a dataframe produced using mutalisk_to_dataframe}
}
\value{
a character vector containing names of metadata columns. If no metadata columns have been added, returns a zero length character vector. (character)
}
\description{
Get a vector of metadata columns from a mutalisk_dataframe.
}
